+++ lib/src/datas/base.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:json_annotation/json_annotation.dart';
-part 'base_response.g.dart';
 
 ///"CurrentRecordCount":1,
 ///"NextPageUrl":"",
@@ -8,8 +7,8 @@ part 'base_response.g.dart';
 ///"PageNumber":1,
 ///"PrevPageUrl":"",
 ///"RecordCount":1}
-@JsonSerializable()
-class BaseResponse {
+
+abstract class BaseResponse {
   @JsonKey(name: "PageNumber")
   final int pageNumber;
   @JsonKey(name: "PageLength")
@@ -27,8 +26,28 @@ class BaseResponse {
     required this.nextPageUrl,
     required this.prevPageUrl,
   });
+}
 
-  factory BaseResponse.fromJson(Map<String, dynamic> json) =>
-      _$BaseResponseFromJson(json);
-  Map<String, dynamic> toJson() => _$BaseResponseToJson(this);
+///basedata for iheath datas
+abstract class BaseData {
+  @JsonKey(name: 'DataID')
+  final String dataId;
+  @JsonKey(name: 'LastChangeTime')
+  final int lastChangeTime;
+  @JsonKey(name: 'MDate')
+  final int mDate;
+  @JsonKey(name: 'Note')
+  final String note;
+  @JsonKey(name: 'TimeZone')
+  final String timeZone;
+  @JsonKey(name: 'DataSource')
+  final String dataSource;
+  BaseData({
+    required this.dataId,
+    required this.lastChangeTime,
+    required this.mDate,
+    required this.note,
+    required this.timeZone,
+    required this.dataSource,
+  });
 }