+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* iheath initial releases with oauth and bp read 
+
+## 0.0.2
+
+* breaking changes
+* added hr,bp,weight model